+++ database/migrations/2020_07_09_120440_create_user_comment_jawaban_table.php
@@ -16,9 +16,9 @@ class CreateUserCommentJawabanTable extends Migration
         Schema::create('user_comment_jawaban', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('isi');
-            $table->int('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->int('jawaban_id');
+            $table->unsignedBigInteger('jawaban_id');
             $table->foreign('jawaban_id')->references('id')->on('jawaban');
             $table->timestamps();
         });
